Separate db and metrics into their own files  - Attach them to flask app later
