+++ chord_drs/db.py
@@ -1,0 +1,5 @@
+from flask_sqlalchemy import SQLAlchemy
+
+__all__ = ["db"]
+
+db = SQLAlchemy()
